+++ modules/updater/updater.py
@@ -1,6 +1,5 @@
 import os
 
-import modules.containers as containers
 from modules.data_collector import set_filename
 from modules.selenium_utils import scrape, setup_webdriver
 from modules.websites import search_links
